+++ src/shared/lib/pins/pins.h
@@ -46,7 +46,7 @@ bool digitalReadPin(Pin_t pin);
 void pinMode(uint8_t pin, uint8_t mode);
 void setupADC(void);
 void tickAnalog(void);
-int analogRead(uint8_t pin);
+uint16_t analogRead(uint8_t pin);
 void stopReading(void);
 void setUpValidPins(void);
 void setUpAnalogDigitalPin(Pin_t button, uint8_t pin, uint16_t threshold);
